// migrations/snapshots/49500c5c0f9968ae56364d302d5d7271864096996e16837ff8e17547eb43c5ad/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgTargetQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Organization: {
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['output'];
      readonly deletedAt: CodecTypes['pg/timestamp-temporal@1']['output'] | null;
      readonly description: CodecTypes['pg/text@1']['output'] | null;
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamp-temporal@1']['output'];
    };
    readonly OrganizationMember: {
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly organizationId: CodecTypes['pg/text@1']['output'];
      readonly role: 'OWNER' | 'MANAGER' | 'MEMBER' | 'GUEST';
      readonly updatedAt: CodecTypes['pg/timestamp-temporal@1']['output'];
      readonly userId: CodecTypes['pg/text@1']['output'];
    };
    readonly User: {
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['output'];
      readonly deletedAt: CodecTypes['pg/timestamp-temporal@1']['output'] | null;
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly password: CodecTypes['pg/text@1']['output'] | null;
      readonly profilePhoto: CodecTypes['pg/text@1']['output'] | null;
      readonly role: 'ADMIN' | 'MANAGER' | 'MEMBER';
      readonly status: 'ACTIVE' | 'BLOCKED';
      readonly updatedAt: CodecTypes['pg/timestamp-temporal@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Organization: {
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['input'];
      readonly deletedAt: CodecTypes['pg/timestamp-temporal@1']['input'] | null;
      readonly description: CodecTypes['pg/text@1']['input'] | null;
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamp-temporal@1']['input'];
    };
    readonly OrganizationMember: {
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly organizationId: CodecTypes['pg/text@1']['input'];
      readonly role: 'OWNER' | 'MANAGER' | 'MEMBER' | 'GUEST';
      readonly updatedAt: CodecTypes['pg/timestamp-temporal@1']['input'];
      readonly userId: CodecTypes['pg/text@1']['input'];
    };
    readonly User: {
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['input'];
      readonly deletedAt: CodecTypes['pg/timestamp-temporal@1']['input'] | null;
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly password: CodecTypes['pg/text@1']['input'] | null;
      readonly profilePhoto: CodecTypes['pg/text@1']['input'] | null;
      readonly role: 'ADMIN' | 'MANAGER' | 'MEMBER';
      readonly status: 'ACTIVE' | 'BLOCKED';
      readonly updatedAt: CodecTypes['pg/timestamp-temporal@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly Organization: {
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['output'];
      readonly deletedAt: CodecTypes['pg/timestamp-temporal@1']['output'] | null;
      readonly description: CodecTypes['pg/text@1']['output'] | null;
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamp-temporal@1']['output'];
    };
    readonly OrganizationMember: {
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly organizationId: CodecTypes['pg/text@1']['output'];
      readonly role: 'OWNER' | 'MANAGER' | 'MEMBER' | 'GUEST';
      readonly updatedAt: CodecTypes['pg/timestamp-temporal@1']['output'];
      readonly userId: CodecTypes['pg/text@1']['output'];
    };
    readonly User: {
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['output'];
      readonly deletedAt: CodecTypes['pg/timestamp-temporal@1']['output'] | null;
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly password: CodecTypes['pg/text@1']['output'] | null;
      readonly profilePhoto: CodecTypes['pg/text@1']['output'] | null;
      readonly role: 'ADMIN' | 'MANAGER' | 'MEMBER';
      readonly status: 'ACTIVE' | 'BLOCKED';
      readonly updatedAt: CodecTypes['pg/timestamp-temporal@1']['output'];
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly Organization: {
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['input'];
      readonly deletedAt: CodecTypes['pg/timestamp-temporal@1']['input'] | null;
      readonly description: CodecTypes['pg/text@1']['input'] | null;
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamp-temporal@1']['input'];
    };
    readonly OrganizationMember: {
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly organizationId: CodecTypes['pg/text@1']['input'];
      readonly role: 'OWNER' | 'MANAGER' | 'MEMBER' | 'GUEST';
      readonly updatedAt: CodecTypes['pg/timestamp-temporal@1']['input'];
      readonly userId: CodecTypes['pg/text@1']['input'];
    };
    readonly User: {
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['input'];
      readonly deletedAt: CodecTypes['pg/timestamp-temporal@1']['input'] | null;
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly password: CodecTypes['pg/text@1']['input'] | null;
      readonly profilePhoto: CodecTypes['pg/text@1']['input'] | null;
      readonly role: 'ADMIN' | 'MANAGER' | 'MEMBER';
      readonly status: 'ACTIVE' | 'BLOCKED';
      readonly updatedAt: CodecTypes['pg/timestamp-temporal@1']['input'];
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;